[refact] 주문 상세 페이지 리팩토링
- 주문 정보 컴포넌트 분리
