+++ frontend/src/pages/user/PurchasePage.tsx
@@ -14,6 +14,7 @@ import CenterWrapper from '@components/atoms/CenterWrapper';
 import Heading from '@components/atoms/Heading';
 import Button from '@components/atoms/Button';
 import ItemPreviewContainer from '@components/pageComponents/CartOrderPage/CartOrderPreview';
+import DeliveryInformation from '@components/pageComponents/UserPage/DeliveryInformation';
 
 interface LocationState {
   state: { productId: string; isDirectPurchase: boolean; quantity: number };
@@ -23,7 +24,7 @@ const PurchasePage = () => {
   const { state } = useLocation() as LocationState;
   const userData = useAppSelector(state => state.user.userData, shallowEqual);
 
-  const { userInfo, userInfoMissing, isLoading } = useFetchUserInfo(userData?._id);
+  const { userInfoMissing, isLoading } = useFetchUserInfo(userData?._id);
   const { orderHandler, setDirectPurchaseItem, orderItemList, orderTotal } = useOrderHandler();
 
   useEffect(() => {
@@ -55,46 +56,7 @@ const PurchasePage = () => {
           </Heading>
           <div className="grid md:grid-cols-[1fr_25%] md:gap-[30px] lg:gap-[60px]">
             <div>
-              <section className="mb-[40px] border-b-[1px] pb-3">
-                <h2 className="font-semibold text-[20px] mb-[10px]">배송지</h2>
-                {userInfoMissing ? (
-                  <span className="text-[red]">
-                    배송지 정보가 없습니다. 프로필에서 배송지 정보를 업데이트 해주세요.
-                  </span>
-                ) : (
-                  <>
-                    <div className="flex mb-2">
-                      <span className="block mr-4 w-[72px]">주소</span>
-                      <span>{userInfo?.address}</span>
-                    </div>
-                    <div className="flex mb-2">
-                      <span className="block mr-4 w-[72px]">우편번호</span>
-                      <span>{userInfo?.zipCode}</span>
-                    </div>
-                  </>
-                )}
-              </section>
-              <section className="mb-[40px] border-b-[1px] pb-3">
-                <h2 className="font-semibold text-[20px] mb-[10px]">주문자</h2>
-                <div className="flex mb-2">
-                  <span className=" block mr-4 w-[72px]">이름</span>
-                  <span>{userData?.name}</span>
-                </div>
-
-                <div className="flex mb-2">
-                  <span className=" block mr-4 w-[72px]">이메일</span>
-                  <span>{userData?.email}</span>
-                </div>
-
-                <div className="flex mb-2">
-                  <span className="block mr-4 w-[72px]">휴대전화</span>
-
-                  <span className={`${userInfo?.phoneNumber ? 'text-black' : 'text-[red]'}`}>
-                    {userInfo?.phoneNumber ||
-                      '휴대전화 정보가 없습니다. 프로필에서 정보를 업데이트 해주세요.'}
-                  </span>
-                </div>
-              </section>
+              <DeliveryInformation />
               <ItemPreviewContainer itemList={orderItemList} isPurchasePage />
             </div>
 
@@ -123,7 +85,7 @@ const PurchasePage = () => {
                 size="full"
                 hovercolor="default"
                 disabled={userInfoMissing || !orderTotal}
-                onClick={() => orderHandler(state.isDirectPurchase)}
+                onClick={() => orderHandler(state?.isDirectPurchase)}
               >
                 결제하기
               </Button>
